Detect monitors by size to load correct autorandr
My binding `$mod+n $mod+Shift+d`, which attempts to force awaken
my Dell monitor and then spread my desktop across the two displays,
wasn't working reliably this morning.

I saw that the `autorandr docked` config was referencing a display
(DP-0) which `xrandr -q` was reporting as disconnected. DP-0 in the
docked config intends to reference the laptop display, which as of
this moment is actually being detected by xrandr as DP-2, not DP-0.

Now we detect the monitors' display values by finding their physical
sizes in the `xrandr q` output, and then load the correct autorandr
profile.

----

For this to work requires having a bunch of different autorandr
profiles. (I tried ditching autorandr for raw xrandr but autorandr
seems to poke the Dell awake in a way that xrandr does not.)

I've added convenience mobile/docked toggles to this command to
mimic the `autorandr {docked,mobile}' interaction which I'm losing.