Use i3-nagbar to present errors
No clue where i3's log is. Just use the nagbar to inform me that
something didn't work as planned.

----

Also, make sure the autorandr profile we try to load corresponds
to one which I've actually saved! This previously failed because
I had a saved profile in autorandr named

    docked-dp1,dp2

but was trying to load

    docked-dp-1,dp-2

The display names from `xrandr` have a hyphen in them which I
needed to accomodate somehow.

I fixed it within autorandr by re-saving my config to be
named

    docked:dp-1,dp-2

(i.e. allow hyphens, and change the delimiter to a colon so the
hyphen isn't overloaded)

Tweak the script a little to handle the new delimiter.