Add keybindings to resize windows
These are a little weird because they're not analogous to tmux.

Tmux: {h,j,k,l} mean "push the edge in this direction".

For example:

    +------------|------------+
    |            |            |
    |     1      |     2      |
    |            |            |
    |            |            |
    +------------|------------+

In tmux, `C-n M-Right` means **different things** depending on where your
active focus is:

* focus in 1: "push my edge [with 2] rightwards" - i.e. "grow my width"
* focus in 2: "push my edge [with 1] rightwards" - i.e. "shrink my width"

The effect for the user is the **same** though, which is that 1 gets
wider and 2 gets narrower. **Different things result in the same effect.**

In i3, there's no accounting for relative position with regards to other
windows, so, `resize grow width` always does the **same thing**, which
depending on relative positioning of windows ends up **feeling different**.

Regardless of where your focus is, `S-n L` always means:

* focus in 1: "make me a little wider, thereby shrinking 2"
* focus in 2: "make me a little wider, thereby shrinking 1"

This is really weird if you're thinking about the edge between the windows
instead of the width of the windows. The edge moves in opposite directions
depending on which window you have selected:

`S-n L` with focus in "1":

    +------------------|------+
    |                  |      |
    |    [1]           |  2   |
    |                  |      |
    |                  |      |
    +------------------|------+

`S-n L` with focus in "2":

    +-----|-------------------+
    |     |                   |
    |  1  |           [2]     |
    |     |                   |
    |     |                   |
    +-----|-------------------+

Ultimately I want to think of the edge between windows, not the widths
(or heights) of windows.

So, add keybindings which reverse the grow/shrink for the times when I try to
move an edge left (right, up, down) and it "wrongly" moves right (left, down, up).
Mnemonic: "oh, it's the ALTernative grow/shrink that I wanted, damnit"