Always put arandr on ws8 on built-in display
I haven't bothered with autorandr to automatically detect
display changes. Instead I've been just manually starting
`Arandr` and deactivating external displays before unplugging them.

Sometimes I have `Arandr` running on a workspace which is showing
on an external display. When this happens, it's a pain to get
`Arandr` to show up for me.

This commit just ensures that `Arandr` is always running on
workspace 8 as soon as i3wm starts or restarts. I still have
the manual step to enable/disable the display, but at least now
I can easily find the program which helps me do so.