Try to start dbus and fix xmodmap at start
This doesn't actually work, and I sort of give up for now.

The new bin/i3-set-inputs properly does start ibus, but it is
unable to do the xmodmap stuff. I don't know why.

To fix the situation where my i3 config is using a modifier key that
my keyboard can't actualy send, I've plumbed in

    s-n $

which is a binding for

   $mod-enter

i.e. it starts a terminal with tmux in it.

So now:

1. log in
2. s-n $
3. source ~/.xinputrc
4. profit?

----

Some day in the future I'll be smart enough to figure this out.